moving bit mask to new file comments
