+++ Assets/Mirage/Runtime/Serialization/BitMask.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e8ce6cd3bc8829743bc1b67cf937135d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
